load function and refactor
