+++ frontend/src/routes/learn/+page.js
@@ -1,0 +1,5 @@
+import { load_word } from "$lib/api/learn"
+
+export async function load() {
+    return await load_word()
+}
